@@ -4,7 +4,7 @@ import setuptools
 
 setuptools.setup(
     name="bntransformer",
-    version="1.0",
+    version="2.0.0",
     author="Sagor Sarker",
     author_email="sagorhem3532@gmail.com",
     description="bengali transformer bengali language processing state of the art transformer",
@@ -18,8 +18,9 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
-    python_requires='>=3.5',
+    python_requires='>=3.6',
     install_requires=[
-        "transformers",
+        "transformers==4.6.1",
+        "sentencepiece",
     ],
 )
