@@ -1,1 +1,9 @@
-__version__="1.0"
+__version__="2.0.0"
+
+from bntransformer.pipeline import (
+    BanglaQA,
+    BanglaNER,
+    BanglaMaskGeneration,
+    BanglaTranslation,
+)
+from bntransformer.tokenizer import BanglaTokenizer
